==> 00 Gather Data.ipynb <==
import pandas as pd
import numpy as np
import os

df_c = pd.read_csv('./data/challengers/atp_matches_qual_chall_1978.csv')
df_a = pd.read_csv('./data/atp/atp_matches_1978.csv')
df_p = pd.read_csv('./data/atp_players.csv')

df_c['winner_name'].head()

df_c.columns

df_c.shape

df_c[df_c['tourney_level'] == 'C']

df_c1 = df_c[['tourney_id', 'tourney_name', 'surface', 'tourney_level', 'winner_id', 'loser_id']]

df_c['tourney_id'].unique(), len(df_c['tourney_id'].unique())

df_c['tourney_name'].unique(), len(df_c['tourney_name'].unique())

df_c['winner_name'].unique(), len(df_c['winner_name'].unique())

df_c['winner_id'].unique(), len(df_c['winner_id'].unique())

df_a

df_a.columns

df_a.columns == df_c.columns

df_a['tourney_name'].unique(), len(df_a['tourney_name'].unique())

df_a['winner_id'].unique(), len(df_a['winner_id'].unique())

df_a['winner_name'].nunique()

df_a['winner_name'].unique(), len(df_a['winner_name'].unique())

df_p.columns

df_p[df_p['player_id'] == 100437]

df_a[df_a['winner_id'] == 100437]['winner_name']

### Combine all Challengers data

path = os.listdir('./data/challengers/') # getting path
df_cha_all = pd.DataFrame(list())
df_cha_all.to_csv('./data/challengers/all.csv') # create an empty csv

for i in path:
    if i != 'all.csv': # make sure it does not loop back to All.csv
        df_cha_all = pd.read_csv('./data/challengers/all.csv')
        df_cha_temp = pd.read_csv('./data/challengers/'+i)
        pd.concat([df_cha_all, df_cha_temp]).to_csv('./data/challengers/all.csv', index=False) # concat to All.csv

df_cha_all = pd.read_csv('./data/challengers/all.csv')
df_cha_all.head(1)

===

### Combine all pro tour data

path = os.listdir('./data/atp/') # getting path
df_atp_all = pd.DataFrame(list())
df_atp_all.to_csv('./data/atp/all.csv') # create an empty csv

for i in path:
    if i != 'all.csv': # make sure it does not loop back to All.csv
        df_atp_all = pd.read_csv('./data/atp/all.csv')
        df_atp_temp = pd.read_csv('./data/atp/'+i)
        pd.concat([df_atp_all, df_atp_temp]).to_csv('./data/atp/all.csv', index=False) # concat to All.csv

df_atp_all = pd.read_csv('./data/atp/all.csv')
df_atp_all.head(1)

df_atp_all.columns

df_atp_all.isnull().sum()

df = df_atp_all

cols = ['w_ace', 'w_df',
       'w_svpt', 'w_1stIn', 'w_1stWon', 'w_2ndWon', 'w_SvGms', 'w_bpSaved',
       'w_bpFaced', 'l_ace', 'l_df', 'l_svpt', 'l_1stIn', 'l_1stWon',
       'l_2ndWon', 'l_SvGms', 'l_bpSaved', 'l_bpFaced']

def chk_df(l, funct, opr):
# this function will get any list, and function and operation
# then create Bitwise Operators for within the list
    a = ''
    for x in range(len(l)):
        if x == len(l)-1:
            a = (a + f'df[\'{l[x]}\'].{funct}')
        else:
            a = (a + f'df[\'{l[x]}\'].{funct} {opr} ')

    a = f'df[{a}]'
    print(a)

chk_df(cols, 'isnull()', '|')

df[df['w_ace'].isnull() | df['w_df'].isnull() | df['w_svpt'].isnull() | df['w_1stIn'].isnull() | df['w_1stWon'].isnull() | df['w_2ndWon'].isnull() | df['w_SvGms'].isnull() | df['w_bpSaved'].isnull() | df['w_bpFaced'].isnull() | df['l_ace'].isnull() | df['l_df'].isnull() | df['l_svpt'].isnull() | df['l_1stIn'].isnull() | df['l_1stWon'].isnull() | df['l_2ndWon'].isnull() | df['l_SvGms'].isnull() | df['l_bpSaved'].isnull() | df['l_bpFaced'].isnull()]
