==> 01 pro_lost.ipynb <==
import pandas as pd

df = pd.read_csv('./data/atp/all.csv')

df.drop(columns='Unnamed: 0', axis=1, inplace=True)

df['loser_name'] = df['loser_name'].str.strip()

df.head()

df['tourney_name'].value_counts()

df['tourney_name'].str.startswith('Davis').value_counts()

df[df['tourney_name'].str.contains('Davis')==False]

df = df[df['tourney_name'].str.contains('Davis')==False]

df.columns

df.drop(columns=['tourney_id','draw_size', 'tourney_level', 'tourney_date', 'winner_seed', 'winner_entry', 'match_num', 'winner_hand', 'winner_ht', 'winner_ioc', 'winner_age', 'winner_id', 'winner_name', 'loser_seed', 'loser_entry', 'loser_hand',
       'loser_ht', 'loser_ioc', 'loser_age', 'score', 'best_of', 'round',
       'minutes', 'w_ace', 'w_df', 'w_svpt', 'w_1stIn', 'w_1stWon', 'w_2ndWon',
       'w_SvGms', 'w_bpSaved', 'w_bpFaced', 'l_ace', 'l_df', 'l_svpt',
       'l_1stIn', 'l_1stWon', 'l_2ndWon', 'l_SvGms', 'l_bpSaved', 'l_bpFaced',
       'winner_rank', 'winner_rank_points', 'loser_rank', 'loser_rank_points'], axis=1, inplace=True)

df_dummies = pd.get_dummies(df, columns=['tourney_name', 'surface'])

df_dummies = df_dummies.sort_values('loser_name')

df_dummies.head()

df_dummies.columns

x = {}
for i in df_dummies.columns:
    if i != 'loser_name' and i != 'loser_id':
        x[i] = ['sum','mean']

df_dummies = df_dummies.groupby('loser_name').agg(x)

df_dummies.head()

df_dummies.drop(index='', inplace=True)

# split columns name sum/mean, into 2 separate columns

df_final = pd.DataFrame()

for x, y in df_dummies.columns:
    df_final[x+" "+y] = df_dummies[x][y]

df_final.columns

# rename columns

for c in df_final.columns:
    if 'tourney_name_' in c:
        c2 = c[13:]
        df_final = df_final.rename(columns={c:c2})

df_final.columns

for x, y in df_dummies.columns:
    total = 0
    avg = 0
    if y == 'sum':
        total += df_dummies[x][y]
    else:
        avg = (df_dummies[x][y]+avg)/2
    
    df_dummies['total_win'] = total
    df_dummies['average_win'] = avg

df_final

df_final.to_csv('./data/atp/lost.csv')

df_final = pd.read_csv('./data/atp/lost.csv')

df_final.head()

df_final.shape
