==> 02 Cha Win Lost Total.ipynb <==
import pandas as pd

df_cw = pd.read_csv('./data/challengers/win.csv')
df_cl = pd.read_csv('./data/challengers/lost.csv')

# change all '_name' columns to 'player_name'
df_cw = df_cw.rename(columns={'winner_name': 'player_name'})
df_cl = df_cl.rename(columns={'loser_name': 'player_name'})

### Total Tour Win

df_cw.shape, df_cl.shape

df_a_total = pd.DataFrame() # create empty dataframe

df_a_total['player_name'] = df_cw['player_name']

for i in df_cw.columns:
    if ('surface' not in i) & ('sum' in i):
        df_a_total[i] = df_cw[i]

df_a_total['win'] = df_a_total.sum(axis=1)

df_final = pd.DataFrame()

df_final['player_name'] = df_cl['player_name']

df_final = pd.merge(df_final, df_a_total[['player_name', 'win']], on='player_name', how='left')

### Total Tour Lost

df_b_total = pd.DataFrame() # create empty dataframe

df_b_total['player_name'] = df_cl['player_name']

for i in df_cl.columns:
        if ('surface' not in i) and ('sum' in i):
            df_b_total[i] = df_cl[i]

df_b_total['lost'] = df_b_total.sum(axis=1)

df_final = pd.merge(df_final, df_b_total[['player_name', 'lost']], on='player_name', how='left')

### Total Match played and avg win

df_final

df_final.fillna(0, inplace=True)

df_final.isnull().sum()

df_final['match_played'] = df_final['win'] + df_final['lost']

df_final['total_win_percentage'] = df_final['win']/df_final['match_played']

df_final['total_lost_percentage'] = df_final['lost']/df_final['match_played']

df_final[df_final['player_name'] == 'Roger Federer']

df_final.to_csv('./data/challengers/Total.csv')

df_final = pd.read_csv('./data/challengers/Total.csv')

df_final
