==> 02 surface win lost.ipynb <==
import pandas as pd

df_aw = pd.read_csv('./data/atp/win.csv') # atp tour wins
df_al = pd.read_csv('./data/atp/lost.csv') # atp tour losts
df_cw = pd.read_csv('./data/challengers/win.csv') # challenger tour wins
df_cl = pd.read_csv('./data/challengers/lost.csv') # cahllenger tour lost

# change all '_name' columns to 'player_name'
df_aw = df_aw.rename(columns={'winner_name': 'player_name'})
df_al = df_al.rename(columns={'loser_name': 'player_name'})
df_cw = df_al.rename(columns={'winner_name': 'player_name'})
df_cl = df_cl.rename(columns={'loser_name': 'player_name'})

### Combine all atp tour surface win/lost

df_aw.head(1)

df_al.head(1)

df_aw.shape, df_al.shape

df_a_surface = pd.DataFrame() # create empty dataframe

df_a_surface['player_name'] = df_al['player_name']

for i in df_al.columns:
        if 'surface' in i:
            df_a_surface[i+'_lost'] = df_al[i]

for i in df_aw.columns:
        if 'surface' in i:
            df_a_surface[i+'_win'] = df_aw[i]

df_a_surface.shape

df_a_surface.columns

df_a_surface = df_a_surface.rename(columns={
    'surface_Carpet sum_lost': 'surface_Carpet_lost',
    'surface_Carpet mean_lost': 'surface_Carpet_lost_avg',
    'surface_Clay sum_lost': 'surface_Clay_lost',
    'surface_Clay mean_lost': 'surface_Clay_lost_avg',
    'surface_Grass sum_lost': 'surface_Grass_lost',
    'surface_Grass mean_lost': 'surface_Grass_lost_avg',
    'surface_Hard sum_lost': 'surface_Hard_lost',
    'surface_Hard mean_lost': 'surface_Hard_lost_avg',
    
    'surface_Carpet sum_win': 'surface_Carpet_win',
    'surface_Carpet mean_win': 'surface_Carpet_win_avg',
    'surface_Clay sum_win': 'surface_Clay_win',
    'surface_Clay mean_win': 'surface_Clay_win_avg',
    'surface_Grass sum_win': 'surface_Grass_win',
    'surface_Grass mean_win': 'surface_Grass_win_avg',
    'surface_Hard sum_win': 'surface_Hard_win',
    'surface_Hard mean_win': 'surface_Hard_win_avg'
})

df_a_surface.fillna('0', inplace=True)

### Combine all Challenger tour surface win/lost

df_cw.head(1)

df_cl.head(1)

df_cw.shape, df_cl.shape

df_c_surface = pd.DataFrame() # create empty dataframe

df_c_surface['player_name'] = df_cl['player_name']

for i in df_cl.columns:
        if 'surface' in i:
            df_c_surface[i+'_lost'] = df_cl[i]

for i in df_cw.columns:
        if 'surface' in i:
            df_c_surface[i+'_win'] = df_cw[i]

df_c_surface.shape

df_c_surface = df_c_surface.rename(columns={
    'surface_Carpet sum_lost': 'surface_Carpet_lost',
    'surface_Carpet mean_lost': 'surface_Carpet_lost_avg',
    'surface_Clay sum_lost': 'surface_Clay_lost',
    'surface_Clay mean_lost': 'surface_Clay_lost_avg',
    'surface_Grass sum_lost': 'surface_Grass_lost',
    'surface_Grass mean_lost': 'surface_Grass_lost_avg',
    'surface_Hard sum_lost': 'surface_Hard_lost',
    'surface_Hard mean_lost': 'surface_Hard_lost_avg',
    
    'surface_Carpet sum_win': 'surface_Carpet_win',
    'surface_Carpet mean_win': 'surface_Carpet_win_avg',
    'surface_Clay sum_win': 'surface_Clay_win',
    'surface_Clay mean_win': 'surface_Clay_win_avg',
    'surface_Grass sum_win': 'surface_Grass_win',
    'surface_Grass mean_win': 'surface_Grass_win_avg',
    'surface_Hard sum_win': 'surface_Hard_win',
    'surface_Hard mean_win': 'surface_Hard_win_avg'
})

df_c_surface.fillna(0, inplace=True)
